data.ts 분리작업
